Replace floor tutorial arrow with VFX arrow (check DMs)
Tutorial text must only disappear once the objective has been done, otherwise a player might miss it and not know what to do.

For the first farmer a player places down (for the tutorial), the wheat must only take 5 seconds to produce so they aren't waiting, and reverts back to normal time after picking it up. (takes 20 normally)

Remove the F to add friend interaction on other players (but of course we keep notification after gifting)